requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
playwright==1.40.0
pandas==2.1.3
numpy==1.25.2
python-dateutil==2.8.2
PyYAML==6.0.1
Jinja2==3.1.2
Flask==3.0.0
schedule==1.2.0
fake-useragent==1.4.0
feedparser==6.0.10
nltk==3.8.1
textstat==0.7.3
colorama==0.4.6
tqdm==4.66.1
pytest==7.4.3
pytest-cov==4.1.0